# packages/api-contract/openapi-v1.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multicountry Pesticide KG API",
    "version": "1.1"
  },
  "paths": {
...
    "/api/v1/entities/{node_id}": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Entity",
        "operationId": "entity_api_v1_entities__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EntityData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/neighborhood": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Neighborhood",
        "operationId": "neighborhood_api_v1_graph_neighborhood_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2,
              "minimum": 1,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GraphData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/path": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Path",
        "operationId": "graph_path_api_v1_graph_path_get",
        "parameters": [
          {
            "name": "start_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Start Id"
            }
          },
          {
            "name": "end_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "End Id"
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 3,
              "title": "Max Depth"
            }
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GraphData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EdgeData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "start_id": {
            "type": "string",
            "title": "Start Id"
          },
          "predicate": {
            "type": "string",
            "title": "Predicate"
          },
          "end_id": {
            "type": "string",
            "title": "End Id"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "start_id",
          "predicate",
          "end_id",
          "jurisdiction",
          "source_record_id",
          "source_url",
          "properties"
        ],
        "title": "EdgeData"
      },
      "EntityData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label_original": {
            "type": "string",
            "title": "Label Original"
          },
          "label_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label En"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "label_original",
          "label_en",
          "type",
          "jurisdiction",
          "source_record_id",
          "source_url",
          "properties"
        ],
        "title": "EntityData"
      },
...
      "Envelope_EntityData_": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "$ref": "#/components/schemas/EntityData"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[EntityData]"
      },
      "Envelope_GraphData_": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "$ref": "#/components/schemas/GraphData"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[GraphData]"
      },
...
      "ErrorDescriptor": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDescriptor"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptor"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "GraphData": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/EntityData"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeData"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphData"
      },
...
    }
  }
}